+++ BackEnd/IdeaManager-API/src/database/migrations/1600398157932-CreateTableUser.ts
@@ -1,0 +1,30 @@
+import { MigrationInterface, QueryRunner, Table } from "typeorm";
+
+export class CreateTableUser1600398157932 implements MigrationInterface {
+  public async up(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.createTable(
+      new Table({
+        name: "users",
+        columns: [
+          {
+            name: "id",
+            type: "uuid",
+            isPrimary: true,
+            generationStrategy: "uuid",
+            default: "uuid_generate_v4()",
+          }, //: string;
+          { name: "name", type: "varchar" }, //: string;
+          { name: "alias", type: "varchar" }, //:string;
+          { name: "email", type: "varchar" }, //: string;
+          { name: "password", type: "varchar" }, //: string;
+          { name: "created_at", type: "timestamp", default: "now()" },
+          { name: "updated_at", type: "timestamp", default: "now()" },
+        ],
+      })
+    );
+  }
+
+  public async down(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.dropTable("users");
+  }
+}