+++ BackEnd/IdeaManager-API/src/models/Users.ts
@@ -1,0 +1,8 @@
+class Users {
+  id: string;
+  name: string;
+  alias:string;
+  email: string;
+  password: string;
+  
+}